core/requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.9.1
bleach==5.0.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
deepl==1.22.0
Django==5.2.4
django-appconf==1.1.0
django-bleach==3.1.0
django-ckeditor-5==0.2.18
django-cleanup==9.0.0
django-compressor==4.5.1
django-crispy-forms==2.4
django-imagekit==5.0.0
django-js-asset==3.1.2
django-modeltranslation==0.19.16
django-ratelimit==4.1.0
django-widget-tweaks==1.5.0
idna==3.10
pilkit==3.0
pillow==11.3.0
psycopg2-binary==2.9.10
pycparser==2.22
python-decouple==3.8
rcssmin==1.1.2
requests==2.32.4
rjsmin==1.2.2
six==1.17.0
sqlparse==0.5.3
tinycss2==1.1.1
tzdata==2025.2
urllib3==2.5.0
webencodings==0.5.1
cloudinary==1.41.0
django-cloudinary-storage==0.3.0
django-axes==6.1.1
django-simple-captcha==0.6.0